Add support for PyPy3
Because why not? :) Seriously, just to support PyPy developers we are
considering running XSnippet API on PyPy.
