+++ .travis.yml
@@ -8,6 +8,8 @@ matrix:
       env: TOXENV=py35
     - python: 3.6
       env: TOXENV=py36
+    - python: pypy3
+      env: TOXENV=pypy
     - python: 3.6
       env: TOXENV=openapi
     - python: 3.6
